fix sort FK column labels
